backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-cloud-language==2.13.1
google-cloud-texttospeech==2.17.1
google-auth==2.23.4
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.4
PyYAML==6.0.1